requirements.txt
Django==2.2.4
django-admin-sortable==2.1.7
django-environ==0.4.5
django-modeladmin-reorder==0.2
django-oauth-toolkit==1.2.0
djangorestframework==3.10.2
gunicorn==19.9.0
psycopg2-binary==2.8.3
whitenoise==4.1.3

-e git+https://bitbucket.org/hub9/django-hybrid-router.git#egg=django-hybrid-router
